requirements.txt
python-dotenv
colorama
requests
google-generativeai

# Development and Quality Tools
mypy>=1.8.0
hypothesis>=6.92.0
ruff>=0.1.9
bandit>=1.7.5